Squashed commit of the following:
commit 9637fa20c750f795a249168950249670d96ab7a9
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Fri Feb 12 11:29:30 2021 +0100

    Changed version.

commit 693df00193fc28a96e6e14b432d0b5bef0e63bca
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Fri Feb 12 11:15:13 2021 +0100

    Added full matrix swipe for blastx mode.

commit 25efab5b88e747aab93585b9f6ef0d7bc5944eb4
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Thu Feb 11 16:15:54 2021 +0100

    Fixed warnings.

commit 2ac422da2f5dc7e7fefc35b2985c1a182a805693
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Thu Feb 11 15:15:27 2021 +0100

    Use int traceback for big DPs.

commit 2d95c48d63c43e1f0ab8224401601cd63b60871f
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Wed Feb 10 16:13:23 2021 +0100

    Use int traceback for big DPs.

commit 44d14450449b87e1fa22ccf4bdf843cf73fc8d2b
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Tue Feb 9 18:44:30 2021 +0100

    Don't vectorize for traceback.

commit a56f10f9fa0c75d9924aa0ce8762467d24d0e2c2
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Tue Feb 9 17:00:50 2021 +0100

    Some cleanup.

commit 04f21bee344c90f4a29077213ff5ecad1f4a56a0
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Tue Feb 9 11:30:29 2021 +0100

    Allow setting the gapped filter evalue.

commit 9169e053eb47eb1a2a986037313d5ea62408cb68
Author: Benjamin Buchfink <buchfink@gmail.com>
Date:   Tue Feb 9 11:05:00 2021 +0100

    Fixed config.query_file refs.

+++ src/lib/blast/nlm_linear_algebra.cpp
@@ -36,8 +36,8 @@
 
 /* Documented in nlm_linear_algebra.h. */
 double **
-Nlm_DenseMatrixNew(int nrows,
-                   int ncols)
+Nlm_DenseMatrixNew(size_t nrows,
+                   size_t ncols)
 {
     int i;             /* iteration index */
     double ** mat;     /* the new matrix */
